Move noise-related tests to an independent testing files
